backend/pyproject.toml
[project]
name = "searchmind-backend"
version = "1.0.0"
description = "Document Search with Vector Retrieval & Re-ranking"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.20",
    "fastapi-mcp>=0.1.5",
    # Document Processing
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "pdf2image>=1.17.0",
    "paddleocr>=2.7.0,<3.0",
    "paddlepaddle>=2.6.0,<4.0",
    # Vector Search & Embeddings
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "transformers>=4.40.0",
    # Text Processing
    "langchain-text-splitters>=0.3.0",
    # Utilities
    "pydantic>=2.8.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.0.0",
    "pillow>=10.0.0",
    "numpy>=1.20.0,<2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]
